apps/api/services/pyproject.toml
[tool.poetry]
name = "services"
version = "0.0.0"
description = ""
authors = ["Silvan Melchior"]
license = "MIT"
packages = [{include = "services"}]

[tool.poetry.dependencies]
python = ">=3.9, <4.0"
fastapi = "^0.100.0"
uvicorn = "^0.23.0"
websockets = "^11.0.3"
openai = "^0.27.8"
replicate = "^0.11.0"
text-generation = "^0.6.0"
sentencepiece = "^0.1.99"
multidict = ">=5.0.0,<6.0.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
ipython = "^8.14.0"
black = {extras = ["d"], version = "^23.7.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
